BE: check rights in routes; message hash added

+++ server/routes/roomsRouter.js
@@ -11,9 +11,10 @@ roomsRouter.route('/')
 .get((req, res, next) => {
   RoomModel.find({})
   .then((rooms) => {
+    const userRooms = rooms.filter((r) => r.users.includes(req.user._id));
     res.statusCode = 200;
     res.setHeader('Content-Type', 'application/json');
-    res.json(rooms);
+    res.json(userRooms);
   }, err => next(err))
   .catch(err => next(err))
 })
@@ -34,7 +35,7 @@ roomsRouter.route('/')
   res.end('PUT operation not supported');
 })
 
-.delete((req, res, next) => {
+.delete((req, res, next) => {     //only for admin
   RoomModel.deleteMany({})
   .then((rooms) => {
     res.statusCode = 200;
@@ -47,12 +48,17 @@ roomsRouter.route('/')
 roomsRouter.route('/:roomId')
 .get((req, res, next) => {
   RoomModel.findById(req.params.roomId)
-  //.populate('creator')
-  //.populate('users')
+  .populate('messages')
   .then((room) => {
-    res.statusCode = 200;
-    res.setHeader('Content-Type', 'application/json');
-    res.json(room);
+    if(room.users.includes(req.user._id)){
+      res.statusCode = 200;
+      res.setHeader('Content-Type', 'application/json');
+      res.json(room);
+    } else {
+      err = new Error('Forbidden');
+      err.status = 403;
+      return next(err);
+    }
   }, err => next(err))
   .catch(err => next(err))
 })
@@ -63,27 +69,51 @@ roomsRouter.route('/:roomId')
 })
 
 .put((req, res, next) => {
-  RoomModel.findByIdAndUpdate(req.params.roomId, {$set: req.body}, {new: true})
-    .then((room) => {
+  RoomModel.findById(req.params.roomId)
+  .then((room) => {
+    if(String(room.creator) === String(req.user._id)){
+
+      RoomModel.findByIdAndUpdate(req.params.roomId, {$set: req.body}, {new: true})
+      .then((room) => {
         res.statusCode = 200
         res.setHeader('Content-Type', 'application/json')
         res.json(room)
-    }, (err) => next(err))
-    .catch((err) => next(err))
+      }, (err) => next(err))
+      .catch((err) => next(err))
+
+    } else {
+      err = new Error('Forbidden');
+      err.status = 403;
+      return next(err);
+    }
+  }, err => next(err))
+  .catch(err => next(err))
 })
 
 .delete((req, res, next) => {
-  RoomModel.findByIdAndRemove(req.params.roomId)
+  RoomModel.findById(req.params.roomId)
   .then((room) => {
-    res.statusCode = 200
-    res.setHeader('Content-Type', 'application/json')
-    res.json(room)
-  }, (err) => next(err))
-  .catch((err) => next(err))
+    if(String(room.creator) === String(req.user._id)){
+
+      RoomModel.findByIdAndRemove(req.params.roomId)
+      .then((room) => {
+          res.statusCode = 200
+          res.setHeader('Content-Type', 'application/json')
+          res.json(room)
+      }, (err) => next(err))
+      .catch((err) => next(err))
+
+    } else {
+      err = new Error('Forbidden');
+      err.status = 403;
+      return next(err);
+    }
+  }, err => next(err))
+  .catch(err => next(err))
 })
  
 roomsRouter.route('/:roomId/messages')
-.get((req, res, next) => {
+/* .get((req, res, next) => {
   const {roomId} = req.params;
   RoomModel.findById(roomId)
   .populate('messages.sender')
@@ -99,14 +129,16 @@ roomsRouter.route('/:roomId/messages')
     }
   }, err => next(err))
   .catch(err => next(err))
-})
+}) */
 
 .post((req, res, next) => {
   const {roomId} = req.params;
   RoomModel.findById(roomId)
   .then(room => {
-    if(room){
-      room.messages.push(req.body)
+    if(room && room.users.includes(req.user._id)){
+      req.body.sender = req.user._id;
+      req.body.textHash = crypto.MD5(req.body.text);
+      room.messages.push(req.body);
       room.save()
       .then((room) => {
         RoomModel.findById(room._id)
@@ -117,6 +149,10 @@ roomsRouter.route('/:roomId/messages')
         })
         .catch(err => next(err))
       })
+    } else if(!room.users.includes(req.user._id)){
+      err = new Error('Forbidden');
+      err.status = 403;
+      return next(err);
     } else {
       err = new Error('Room ' + roomId + ' not found')
       err.status = 404
@@ -135,7 +171,7 @@ roomsRouter.route('/:roomId/messages')
   const {roomId} = req.params;
   RoomModel.findById(roomId)
   .then(room => {
-    if(room){
+    if(room && String(room.creator) === String(req.user._id)){
       room.messages = []
       room.save()
       .then((room) => {
@@ -147,6 +183,10 @@ roomsRouter.route('/:roomId/messages')
         })
       })
       .catch(err => next(err))
+    } else if(room.creator !== req.user._id){
+      err = new Error('Forbidden');
+      err.status = 403;
+      return next(err);
     } else {
       err = new Error('Room ' + roomId + ' not found')
       err.status = 404
@@ -157,7 +197,7 @@ roomsRouter.route('/:roomId/messages')
 })
 
 roomsRouter.route('/:roomId/messages/:messageId')
-.get((req, res, next) => {
+.get((req, res, next) => {      //why?
   const {roomId, messageId} = req.params;
   
   RoomModel.findById(roomId)
@@ -186,25 +226,30 @@ roomsRouter.route('/:roomId/messages/:messageId')
   res.end('POST operation not supported')
 })
 
-.put((req, res, next) => { //todo: check rigths
+.put((req, res, next) => {
   const {roomId, messageId} = req.params;
   const {text} = req.body;
-
   RoomModel.findById(roomId)
   .then((room) => {
     const message = room.messages.id(messageId);
-    if(room && message){
+    if(room && message && String(message.sender) === String(req.user._id)){
       message.text = text;
+      message.textHash = crypto.MD5(text);
       room.save()
       .then((room) => {
         RoomModel.findById(room._id)
+        .populate('messages.sender')
         .then((room) => {
           res.statusCode = 200;
           res.setHeader('Content-Type', 'application/json');
-          res.json(message)
+          res.json(room.messages.find((m) => String(m._id) === String(messageId)))
         })
       })
       .catch(err => next(err))
+    } else if(message.sender !== req.user._id){
+      err = new Error('Forbidden');
+      err.status = 403;
+      return next(err);
     } else if(!room) {
       err = new Error('Room ' + roomId + ' not found')
       err.status = 404
@@ -218,13 +263,13 @@ roomsRouter.route('/:roomId/messages/:messageId')
   .catch(err => next(err))
 })
 
-.delete((req, res, next) => { //todo: check rigths
+.delete((req, res, next) => {
   const {roomId, messageId} = req.params;
 
   RoomModel.findById(roomId)
   .then((room) => {
     const message = room.messages.id(messageId);
-    if(room && message){
+    if(room && message && String(message.sender) === String(req.user._id)){
       message.remove()
       room.save()
       .then((room) => {
@@ -235,6 +280,10 @@ roomsRouter.route('/:roomId/messages/:messageId')
           res.json(message);
         })
       })
+    } else if(message.sender !== req.user._id){
+      err = new Error('Forbidden');
+      err.status = 403;
+      return next(err);
     } else if(!room) {
       err = new Error('Room ' + roomId + ' not found')
       err.status = 404
